+++ caten_music/routes/login.py
@@ -6,7 +6,7 @@ from flask_login import login_user, current_user
 
 import datetime
 
-from caten_worship import helper
+from caten_music import helper
 
 
 login_bp = Blueprint("login_bp", __name__,