+++ caten_music/routes/register.py
@@ -4,9 +4,9 @@ from flask import Blueprint, render_template, abort, request, redirect, jsonify,
 from jinja2 import TemplateNotFound
 from flask_login import current_user
 
-from caten_worship import services
-from caten_worship import helper
-from caten_worship import models
+from caten_music import services
+from caten_music import helper
+from caten_music import models
 
 User = models.User
 UserProfile = models.UserProfile
@@ -85,7 +85,7 @@ def register():
         # 寄出帳號啟動 email
         services.send_mail(sender='Sender@domain.com',
                                recipients=[email],
-                               subject='Caten Worship 帳號註冊認證信',
+                               subject='Caten music 帳號註冊認證信',
                                template='activation/verifymail.html',
                                username=username,
                                mail_title="帳號註冊認證信",